rename and change project setup pre-merge with adding_tests
