--- .pre-commit-config.yaml
@@ -1,25 +1,0 @@
-repos:
-    - repo: https://github.com/pycqa/isort
-      rev: 5.10.1
-      hooks:
-          - id: isort
-    - repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v4.2.0
-      hooks:
-          - id: check-docstring-first
-          - id: check-executables-have-shebangs
-          - id: check-merge-conflict
-          - id: check-toml
-          - id: end-of-file-fixer
-          - id: mixed-line-ending
-            args: [--fix=lf]
-          - id: requirements-txt-fixer
-          - id: trailing-whitespace
-    - repo: https://gitlab.com/pycqa/flake8
-      rev: 3.9.2
-      hooks:
-          - id: flake8
-    - repo: https://github.com/psf/black
-      rev: 22.3.0
-      hooks:
-          - id: black
