+++ datashuttle/datashuttle.py
@@ -6,19 +6,19 @@ from typing import Union, cast
 
 import paramiko
 
-import manager.configs as configs
-from manager.utils_mod import rclone_utils, utils
-from manager.utils_mod.decorators import requires_ssh_configs
-from manager.utils_mod.directory_class import Directory
+import datashuttle.configs as configs
+from datashuttle.utils_mod import rclone_utils, utils
+from datashuttle.utils_mod.decorators import requires_ssh_configs
+from datashuttle.utils_mod.directory_class import Directory
 
 # --------------------------------------------------------------------------------------------------------------------
 # Project Manager Class
 # --------------------------------------------------------------------------------------------------------------------
 
 
-class ProjectManager:
+class DataShuttle:
     """
-    Main project manager class for data organisation and transfer in BIDS-style project directory.
+    Main datashuttle class for data organisation and transfer in BIDS-style project directory.
     The expected organisation is a central repository on a remote machine ('remote') that
     contains all project data. This is connected to multiple local machines ('local') which
     each contain a subset of the full project (e.g. machine for electrophysiology collection,
@@ -94,7 +94,7 @@ class ProjectManager:
                 },
             ),
             "imaging": Directory(
-                "imaging_",
+                "imaging",
                 self.cfg["use_imaging"],
             ),
             "histology": Directory(
@@ -278,8 +278,8 @@ class ProjectManager:
         use_histology: bool = True,
     ):
         """
-        Initialise a config file for using the project manager on the local system. Once initialised, these
-        settings will be used each time the project manager is opened.
+        Initialise a config file for using the datashuttle on the local system. Once initialised, these
+        settings will be used each time the datashuttle is opened.
 
         :param local_path:                  path to project dir on local machine
         :param remote_path:                 path to project directory on remote machine. Note this cannot
@@ -329,8 +329,7 @@ class ProjectManager:
 
         self.set_attributes_after_config_load()
         utils.message_user(
-            "Configuration file has been saved and options loaded into the project"
-            " manager."
+            "Configuration file has been saved and options loaded into datashuttle."
         )
 
     def attempt_load_configs(self, prompt_on_fail: bool) -> Union[bool, dict]:
