+++ datashuttle/tests/test_make_dirs.py
@@ -7,7 +7,7 @@ from os.path import join
 import pytest
 import test_utils
 
-from manager.utils_mod import utils
+from datashuttle.utils_mod import utils
 
 
 class TestMakeDirs:
