--- notes.md
@@ -1,18 +1,0 @@
-### Directory Tree Maker
-https://tree.nathanfriend.io
-```
-project_name/
-  raw_data/
-    ephys/
-      mouse/
-        session/
-            behav
-              camera
-    behav
-      mouse/
-        session/
-    microscopy
-      mouse/
-      
-
-```
