--- pyproject.toml
@@ -1,20 +1,0 @@
-[tool.black]
-target-version = ['py36', 'py37', 'py38']
-skip-string-normalization = false
-line-length = 79
-exclude = '''
-(
-  /(
-      \.eggs
-    | \.git
-    | \.hg
-    | \.mypy_cache
-    | \.tox
-    | \.venv
-    | _build
-    | buck-out
-    | build
-    | dist
-  )/
-)
-'''
