--- requirements.txt
@@ -1,7 +1,0 @@
-appdirs==1.4.4
-ftpsync==1.2.2
-paramiko==2.11.0
-pyftpsync==4.0.0
-pytest==7.1.3
-PyYAML==6.0
-setuptools==60.2.0
