--- setup.cfg
@@ -1,16 +1,0 @@
-[flake8]
-# Ignores - https://lintlyci.github.io/Flake8Rules
-# W291 trailing whitespace
-# E731 do not assign a lambda expression, use a def
-# W293 blank line contains whitespace
-# E203  Whitespace before ':'  (sometimes conflicts with black)
-# E501 line too long (84 > 79 characters)  (sometimes too annoying)
-# W503	Line break occurred before a binary operator
-# C901 McCabe complexity test. Would be nice to reenable, but takes work
-ignore = E203,W503,E501,E731,C901,W291,W293,E402,E722
-max-line-length = 79
-max-complexity = 18
-exclude = __init__.py
-
-[isort]
-profile = black
